+++ README.md
@@ -64,6 +64,12 @@ Here are some examples of SVG images generated by this library.
                 alt="Radar thumbnail of Pillbox"
             />
         </td>
+        <td>
+            <img
+                src="./examples/unknown_mini.svg"
+                alt="Radar thumbnail of Unknown Mini"
+            />
+        </td>
     </tr>
 </table>
 
@@ -78,7 +84,7 @@ This project currently has support for drawing the following BZW 2.x object type
 | box         | :white_check_mark: |
 | cone        | :x:                |
 | group       | :white_check_mark: |
-| mesh        | :x:                |
+| mesh        | :white_check_mark: |
 | pyramid     | :white_check_mark: |
 | sphere      | :x:                |
 | teleporter  | :white_check_mark: |
