+++ src/graphics/SVG/Radar/MeshRenderer.php
@@ -13,14 +13,16 @@ use allejo\bzflag\graphics\Common\WorldBoundary;
 use allejo\bzflag\graphics\SVG\ISVGStylable;
 use allejo\bzflag\graphics\SVG\Radar\Styles\MeshStyle;
 use allejo\bzflag\graphics\SVG\SVGStylableUtilities;
+use allejo\bzflag\world\Object\MeshObstacle;
+use SVG\Nodes\Shapes\SVGPolygon;
 use SVG\Nodes\Structures\SVGGroup;
 use SVG\Nodes\SVGNode;
 
 /**
  * @internal
  *
- * @extends ObstacleRenderer<\allejo\bzflag\world\Object\MeshObstacle>
- * @implements ISVGStylable<\allejo\bzflag\world\Object\MeshObstacle>
+ * @extends ObstacleRenderer<MeshObstacle>
+ * @implements ISVGStylable<MeshObstacle>
  */
 class MeshRenderer extends ObstacleRenderer implements ISVGStylable
 {
@@ -57,13 +59,27 @@ class MeshRenderer extends ObstacleRenderer implements ISVGStylable
         return $svg;
     }
 
+    protected function meshToSVGNode(string $class): SVGNode
+    {
+        $svg = new $class();
+
+        foreach ($this->obstacle->getFaces() as $meshFace)
+        {
+            $mesh = new SVGPolygon($meshFace->getVertices());
+            $svg->addChild($mesh);
+        }
+
+        return $svg;
+    }
+
     /**
      * @param null|MeshObstacle $obstacle
      */
     public static function attachBzwAttributes(SVGNode $node, $obstacle): void
     {
+        $node->setAttribute('bzw:type', (string)$obstacle->getObjectType());
         $node->setAttribute('bzw:vertices', json_encode($obstacle->getVertices()));
-    
+        $node->setAttribute('bzw:faces', json_encode($obstacle->getFaces()));
     }
 
     /**
