+++ tests/graphics/fixtures/unknown_mini.bzw
