+++ tests/graphics/fixtures/unknown_mini.rec
